after edit rate to float

+++ database/migrations/2020_05_16_052614_add_rate_to_books.php
@@ -14,7 +14,7 @@ class AddRateToBooks extends Migration
     public function up()
     {
         Schema::table('books', function (Blueprint $table) {
-            $table->unsignedInteger("rate")->nullable();
+            $table->float("rate")->nullable();
         });
     }
